requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.8.2
yfinance==0.2.52
pycoingecko==3.2.0
langchain==0.2.16
langsmith==0.1.116
jinja2==3.1.4
